requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
jinja2>=3.1
pydantic>=2.8
httpx>=0.27
playwright>=1.45
beautifulsoup4>=4.12
lxml>=5.2
readability-lxml>=0.8
sqlmodel>=0.0.21
aiosqlite>=0.20
python-multipart>=0.0.9
python-dotenv>=1.0
jsonschema>=4.23
pyld>=2.0
# dev:
ruff>=0.6
pytest>=8.2
pytest-asyncio>=0.23
